requirements.txt
amqp==1.4.9
anyjson==0.3.3
astroid==2.2.5
autopep8==1.4.4
backcall==0.1.0
billiard==3.3.0.23
celery==3.1.18
certifi==2019.3.9
chardet==3.0.4
decorator==4.4.0
Django==2.2.1
django-environ==0.4.5
django-extensions==2.1.7
django-filter==2.1.0
django-material==1.5.8
django-material-widgets==1.0.0b3
django-slack==5.13.0
django-widget-tweaks==1.4.5
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
freezegun==0.3.12
get==2019.4.13
idna==2.8
ipdb==0.12
ipython==7.5.0
ipython-genutils==0.2.0
isort==4.3.20
jedi==0.13.3
kombu==3.0.37
lazy-object-proxy==1.4.1
Markdown==3.1.1
mccabe==0.6.1
mock==3.0.5
model-mommy==1.6.0
parso==0.4.0
pep8==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
post==2019.4.13
prompt-toolkit==2.0.9
psycopg2==2.8.2
ptyprocess==0.6.0
public==2019.4.13
pycodestyle==2.5.0
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.3.1
pylint-django==2.0.9
pylint-plugin-utils==0.5
python-dateutil==2.8.0
pytz==2019.1
query-string==2019.4.13
redis==2.10.3
request==2019.4.13
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
typed-ast==1.4.0
urllib3==1.25.3
uuid==1.30
wcwidth==0.1.7
wrapt==1.11.1
